Extract `SearchMetadataResource`, separate test suites.
