+++ graylog2-server/src/main/java/org/graylog/plugins/views/ViewsBindings.java
@@ -58,6 +58,7 @@ import org.graylog.plugins.views.search.rest.MessagesResource;
 import org.graylog.plugins.views.search.rest.PivotSeriesFunctionsResource;
 import org.graylog.plugins.views.search.rest.QualifyingViewsResource;
 import org.graylog.plugins.views.search.rest.SavedSearchesResource;
+import org.graylog.plugins.views.search.rest.SearchMetadataResource;
 import org.graylog.plugins.views.search.rest.SearchResource;
 import org.graylog.plugins.views.search.rest.ViewsResource;
 import org.graylog.plugins.views.search.rest.ViewsRestPermissions;
@@ -124,6 +125,7 @@ public class ViewsBindings extends ViewsModule {
         addSystemRestResource(QualifyingViewsResource.class);
         addSystemRestResource(SavedSearchesResource.class);
         addSystemRestResource(SearchResource.class);
+        addSystemRestResource(SearchMetadataResource.class);
         addSystemRestResource(ViewsResource.class);
 
         addPermissions(ViewsRestPermissions.class);
