+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/rest/SearchResource.java
@@ -18,24 +18,17 @@ package org.graylog.plugins.views.search.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Maps;
 import com.google.common.eventbus.EventBus;
 import com.google.common.util.concurrent.Uninterruptibles;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import one.util.streamex.StreamEx;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
 import org.graylog.plugins.views.audit.ViewsAuditEventTypes;
-import org.graylog.plugins.views.search.Parameter;
-import org.graylog.plugins.views.search.Query;
-import org.graylog.plugins.views.search.QueryMetadata;
 import org.graylog.plugins.views.search.Search;
 import org.graylog.plugins.views.search.SearchDomain;
 import org.graylog.plugins.views.search.SearchExecutionGuard;
 import org.graylog.plugins.views.search.SearchJob;
-import org.graylog.plugins.views.search.SearchMetadata;
-import org.graylog.plugins.views.search.db.SearchDbService;
 import org.graylog.plugins.views.search.db.SearchJobService;
 import org.graylog.plugins.views.search.engine.QueryEngine;
 import org.graylog.plugins.views.search.events.SearchJobExecutionEvent;
@@ -51,8 +44,8 @@ import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import javax.validation.constraints.NotNull;
+import javax.ws.rs.Consumes;
 import javax.ws.rs.DefaultValue;
-import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.GET;
 import javax.ws.rs.InternalServerErrorException;
 import javax.ws.rs.NotFoundException;
@@ -66,9 +59,6 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.net.URI;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
@@ -78,38 +68,34 @@ import static com.google.common.base.MoreObjects.firstNonNull;
 @Api(value = "Search")
 @Path("/views/search")
 @Produces(MediaType.APPLICATION_JSON)
+@Consumes({MediaType.APPLICATION_JSON})
 @RequiresAuthentication
 public class SearchResource extends RestResource implements PluginRestResource {
     private static final Logger LOG = LoggerFactory.getLogger(SearchResource.class);
-
     private static final String BASE_PATH = "views/search";
 
-    private final QueryEngine queryEngine;
-    private final SearchDbService searchDbService;
+    private final SearchExecutionGuard executionGuard;
+    private final SearchDomain searchDomain;
     private final SearchJobService searchJobService;
     private final ObjectMapper objectMapper;
     private final PermittedStreams permittedStreams;
-    private final SearchExecutionGuard executionGuard;
-    private final SearchDomain searchDomain;
+    private final QueryEngine queryEngine;
     private final EventBus serverEventBus;
 
-
     @Inject
-    public SearchResource(QueryEngine queryEngine,
-                          SearchDbService searchDbService,
+    public SearchResource(SearchExecutionGuard executionGuard,
+                          SearchDomain searchDomain,
                           SearchJobService searchJobService,
                           ObjectMapper objectMapper,
                           PermittedStreams permittedStreams,
-                          SearchExecutionGuard executionGuard,
-                          SearchDomain searchDomain,
+                          QueryEngine queryEngine,
                           EventBus serverEventBus) {
-        this.queryEngine = queryEngine;
-        this.searchDbService = searchDbService;
+        this.executionGuard = executionGuard;
+        this.searchDomain = searchDomain;
         this.searchJobService = searchJobService;
         this.objectMapper = objectMapper;
         this.permittedStreams = permittedStreams;
-        this.executionGuard = executionGuard;
-        this.searchDomain = searchDomain;
+        this.queryEngine = queryEngine;
         this.serverEventBus = serverEventBus;
     }
 
@@ -117,19 +103,14 @@ public class SearchResource extends RestResource implements PluginRestResource {
     @ApiOperation(value = "Create a search query", response = Search.class, code = 201)
     @AuditEvent(type = ViewsAuditEventTypes.SEARCH_CREATE)
     public Response createSearch(@ApiParam Search search, @Context SearchUser searchUser) {
-        final Optional<Search> previous = Optional.ofNullable(search.id()).flatMap(searchDbService::get);
-        if (!searchUser.isAdmin() && !previous.map(searchUser::owns).orElse(true)) {
-            throw new ForbiddenException("Unable to update search with id <" + search.id() + ">, already exists and user is not permitted to overwrite it.");
-        }
-
         guard(search, searchUser);
 
-        final Search saved = searchDbService.save(search.toBuilder().owner(searchUser.username()).build());
+        final Search saved = searchDomain.saveForUser(search, searchUser);
         if (saved == null || saved.id() == null) {
             return Response.serverError().build();
         }
         LOG.debug("Created new search object {}", saved.id());
-        return Response.created(URI.create(Objects.requireNonNull(saved.id()))).entity(saved).build();
+        return Response.created(URI.create(saved.id())).entity(saved).build();
     }
 
     @GET
@@ -149,13 +130,14 @@ public class SearchResource extends RestResource implements PluginRestResource {
 
     @POST
     @ApiOperation(value = "Execute the referenced search query asynchronously",
-            notes = "Starts a new search, irrespective whether or not another is already running")
+                  notes = "Starts a new search, irrespective whether or not another is already running")
     @Path("{id}/execute")
     @NoAuditEvent("Creating audit event manually in method body.")
     public Response executeQuery(@ApiParam(name = "id") @PathParam("id") String id,
                                  @ApiParam ExecutionState executionState,
                                  @Context SearchUser searchUser) {
-        Search search = getSearch(id, searchUser);
+        Search search = searchDomain.getForUser(id, searchUser)
+                .orElseThrow(() -> new NotFoundException("Search with id " + id + " does not exist"));
 
         search = search.addStreamsToQueriesWithoutStreams(() -> loadAllAllowedStreamsForUser(searchUser));
 
@@ -172,19 +154,6 @@ public class SearchResource extends RestResource implements PluginRestResource {
         return Response.created(URI.create(BASE_PATH + "/status/" + runningSearchJob.getId()))
                 .entity(runningSearchJob)
                 .build();
-    }
-
-    private void postAuditEvent(SearchJob searchJob) {
-        final SearchJobExecutionEvent searchJobExecutionEvent = SearchJobExecutionEvent.create(getCurrentUser(), searchJob, DateTime.now(DateTimeZone.UTC));
-        this.serverEventBus.post(searchJobExecutionEvent);
-    }
-
-    private ImmutableSet<String> loadAllAllowedStreamsForUser(SearchUser searchUser) {
-        return permittedStreams.load(searchUser::canReadStream);
-    }
-
-    private void guard(Search search, SearchUser searchUser) {
-        this.executionGuard.check(search, searchUser::canReadStream);
     }
 
     @POST
@@ -233,25 +202,16 @@ public class SearchResource extends RestResource implements PluginRestResource {
         return searchJob;
     }
 
-    @GET
-    @ApiOperation(value = "Metadata for the given Search object", notes = "Used for already persisted search objects")
-    @Path("metadata/{searchId}")
-    public SearchMetadata metadata(@ApiParam("searchId") @PathParam("searchId") String searchId, @Context SearchUser searchUser) {
-        final Search search = getSearch(searchId, searchUser);
-        return metadataForObject(search);
+    private void guard(Search search, SearchUser searchUser) {
+        this.executionGuard.check(search, searchUser::canReadStream);
     }
 
-    @POST
-    @ApiOperation(value = "Metadata for the posted Search object", notes = "Intended for search objects that aren't yet persisted (e.g. for validation or interactive purposes)")
-    @Path("metadata")
-    @NoAuditEvent("Only returning metadata for given search, not changing any data")
-    public SearchMetadata metadataForObject(@ApiParam @NotNull Search search) {
-        if (search == null) {
-            throw new IllegalArgumentException("Search must not be null.");
-        }
-        final Map<String, QueryMetadata> queryMetadatas = StreamEx.of(search.queries()).toMap(Query::id, query -> queryEngine.parse(search, query));
-        return SearchMetadata.create(queryMetadatas, Maps.uniqueIndex(search.parameters(), Parameter::name));
+    private void postAuditEvent(SearchJob searchJob) {
+        final SearchJobExecutionEvent searchJobExecutionEvent = SearchJobExecutionEvent.create(getCurrentUser(), searchJob, DateTime.now(DateTimeZone.UTC));
+        this.serverEventBus.post(searchJobExecutionEvent);
     }
 
-
+    private ImmutableSet<String> loadAllAllowedStreamsForUser(SearchUser searchUser) {
+        return permittedStreams.load(searchUser::canReadStream);
+    }
 }
