+++ graylog2-server/src/test/java/org/graylog/plugins/views/search/rest/SearchResourceTest.java
@@ -16,7 +16,6 @@
  */
 package org.graylog.plugins.views.search.rest;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.eventbus.EventBus;
 import org.graylog.plugins.views.search.Query;
@@ -24,26 +23,20 @@ import org.graylog.plugins.views.search.Search;
 import org.graylog.plugins.views.search.SearchDomain;
 import org.graylog.plugins.views.search.SearchExecutionGuard;
 import org.graylog.plugins.views.search.SearchJob;
-import org.graylog.plugins.views.search.db.SearchDbService;
 import org.graylog.plugins.views.search.db.SearchJobService;
 import org.graylog.plugins.views.search.engine.QueryEngine;
-import org.graylog.plugins.views.search.events.SearchJobExecutionEvent;
 import org.graylog.plugins.views.search.permissions.SearchUser;
-import org.graylog2.plugin.database.users.User;
 import org.graylog2.shared.bindings.GuiceInjectorHolder;
 import org.graylog2.shared.bindings.providers.ObjectMapperProvider;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-import javax.annotation.Nullable;
 import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.NotFoundException;
-import javax.ws.rs.core.Response;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
@@ -54,8 +47,6 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -66,54 +57,32 @@ public class SearchResourceTest {
     public MockitoRule rule = MockitoJUnit.rule();
 
     @Mock
-    private QueryEngine queryEngine;
-
-    @Mock
-    private SearchDbService searchDbService;
-
-    @Mock
-    private SearchJobService searchJobService;
-
-    @Mock
-    private PermittedStreams permittedStreams;
-
-    @Mock
     private SearchExecutionGuard executionGuard;
 
     @Mock
     private SearchDomain searchDomain;
 
     @Mock
-    private User currentUser;
+    private SearchJobService searchJobService;
 
     @Mock
     private SearchUser searchUser;
 
     @Mock
+    private PermittedStreams permittedStreams;
+
+    @Mock
     private EventBus eventBus;
 
+    @Mock
+    private QueryEngine queryEngine;
+
     private SearchResource searchResource;
-
-    class SearchTestResource extends SearchResource {
-        SearchTestResource(QueryEngine queryEngine, SearchDbService searchDbService, SearchJobService searchJobService, ObjectMapper objectMapper, PermittedStreams streamLoader) {
-            super(queryEngine, searchDbService, searchJobService, objectMapper, streamLoader, executionGuard, searchDomain, eventBus);
-        }
-
-        @Nullable
-        @Override
-        protected User getCurrentUser() {
-            return currentUser;
-        }
-    }
 
     @Before
     public void setUp() throws Exception {
         GuiceInjectorHolder.createInjector(Collections.emptyList());
-
-        when(searchDbService.get(null)).thenThrow(NullPointerException.class);
-        this.searchResource = new SearchTestResource(queryEngine, searchDbService, searchJobService, objectMapperProvider.get(), permittedStreams);
-
-        when(currentUser.getName()).thenReturn("admin");
+        this.searchResource = new SearchResource(executionGuard, searchDomain, searchJobService, objectMapperProvider.get(), permittedStreams, queryEngine, eventBus);
     }
 
     @Test
@@ -123,10 +92,7 @@ public class SearchResourceTest {
 
         this.searchResource.createSearch(search, searchUser);
 
-        final ArgumentCaptor<Search> ownerCaptor = ArgumentCaptor.forClass(Search.class);
-        verify(searchDbService).save(ownerCaptor.capture());
-
-        assertThat(ownerCaptor.getValue().owner()).isEqualTo(Optional.of("eberhard"));
+        verify(searchDomain).saveForUser(any(), eq(searchUser));
     }
 
     @Test
@@ -155,154 +121,6 @@ public class SearchResourceTest {
     }
 
     @Test
-    public void executeQueryAddsCurrentUserAsOwner() {
-        mockCurrentUserName("basti");
-
-        final Search search = mockExistingSearch();
-
-        this.searchResource.executeQuery(search.id(), ExecutionState.empty(), searchUser);
-
-        final ArgumentCaptor<String> usernameCaptor = ArgumentCaptor.forClass(String.class);
-        verify(searchJobService, times(1)).create(eq(search), usernameCaptor.capture());
-
-        assertThat(usernameCaptor.getValue()).isEqualTo("basti");
-    }
-
-    @Test
-    public void executeQueryTriggersEvent() {
-        mockCurrentUserName("basti");
-
-        final Search search = mockExistingSearch();
-
-        final Response response = this.searchResource.executeQuery(search.id(), ExecutionState.empty(), searchUser);
-
-        final ArgumentCaptor<SearchJobExecutionEvent> eventCaptor = ArgumentCaptor.forClass(SearchJobExecutionEvent.class);
-        verify(this.eventBus, times(1)).post(eventCaptor.capture());
-
-        final SearchJobExecutionEvent searchJobExecutionEvent = eventCaptor.getValue();
-        assertThat(searchJobExecutionEvent.user()).isEqualTo(currentUser);
-        assertThat(searchJobExecutionEvent.searchJob()).isEqualTo(response.getEntity());
-    }
-
-    @Test
-    public void guardExceptionPreventsTriggeringEvent() {
-        mockCurrentUserName("basti");
-
-        final Search search = mockExistingSearch();
-        throwGuardExceptionFor(search);
-
-        try {
-            this.searchResource.executeQuery(search.id(), ExecutionState.empty(), searchUser);
-        } catch (ForbiddenException ignored) {}
-
-        verify(this.eventBus, never()).post(any(SearchJobExecutionEvent.class));
-    }
-
-    @Test
-    public void executeSyncJobTriggersEvent() {
-        mockCurrentUserName("peterchen");
-
-        final Search search = mockExistingSearch();
-
-        final Response response = this.searchResource.executeSyncJob(search, 100, searchUser);
-
-        final ArgumentCaptor<SearchJobExecutionEvent> eventCaptor = ArgumentCaptor.forClass(SearchJobExecutionEvent.class);
-        verify(this.eventBus, times(1)).post(eventCaptor.capture());
-
-        final SearchJobExecutionEvent searchJobExecutionEvent = eventCaptor.getValue();
-        assertThat(searchJobExecutionEvent.user()).isEqualTo(currentUser);
-        assertThat(searchJobExecutionEvent.searchJob()).isEqualTo(response.getEntity());
-    }
-
-    @Test
-    public void guardExceptionDuringSyncExecutionPreventsTriggeringEvent() {
-        mockCurrentUserName("basti");
-
-        final Search search = mockExistingSearch();
-        throwGuardExceptionFor(search);
-
-        try {
-            this.searchResource.executeSyncJob(search, 100, searchUser);
-        } catch (ForbiddenException ignored) {}
-
-        verify(this.eventBus, never()).post(any(SearchJobExecutionEvent.class));
-    }
-
-    @Test
-    public void executeSyncJobAddsCurrentUserAsOwner() {
-        mockCurrentUserName("peterchen");
-
-        final Search search = mockExistingSearch();
-
-        this.searchResource.executeSyncJob(search, 100, searchUser);
-
-        final ArgumentCaptor<String> usernameCaptor = ArgumentCaptor.forClass(String.class);
-        verify(searchJobService, times(1)).create(eq(search), usernameCaptor.capture());
-
-        assertThat(usernameCaptor.getValue()).isEqualTo("peterchen");
-    }
-
-    @Test
-    public void executeQueryAppliesExecutionState() {
-        final Search search = mockExistingSearch();
-
-        final ExecutionState.Builder builder = ExecutionState.builder();
-        builder.addAdditionalParameter("foo", 42);
-
-
-        when(searchDbService.get(search.id())).thenReturn(Optional.of(search));
-
-        final ExecutionState executionState = builder.build();
-        this.searchResource.executeQuery(search.id(), executionState, searchUser);
-
-        //noinspection unchecked
-        final ArgumentCaptor<ExecutionState> executionStateCaptor = ArgumentCaptor.forClass(ExecutionState.class);
-        verify(search, times(1)).applyExecutionState(any(), executionStateCaptor.capture());
-
-        assertThat(executionStateCaptor.getValue()).isEqualTo(executionState);
-    }
-
-    @Test
-    public void guardExceptionInAsyncExecutionLeadsTo403() {
-        final Search search = mockExistingSearch();
-
-        throwGuardExceptionFor(search);
-
-        assertThatExceptionOfType(ForbiddenException.class)
-                .isThrownBy(() -> this.searchResource.executeQuery(search.id(), null, searchUser));
-    }
-
-    @Test
-    public void guardExceptionInSynchronousExecutionLeadsTo403() {
-        final Search search = mockExistingSearch();
-
-        throwGuardExceptionFor(search);
-
-        assertThatExceptionOfType(ForbiddenException.class)
-                .isThrownBy(() -> this.searchResource.executeSyncJob(search, 0, searchUser));
-    }
-
-    @Test
-    public void failureToAddDefaultStreamsInAsyncSearchLeadsTo403() {
-        final Search search = mockExistingSearch();
-
-        doThrow(new ForbiddenException()).when(search).addStreamsToQueriesWithoutStreams(any());
-
-        assertThatExceptionOfType(ForbiddenException.class)
-                .isThrownBy(() -> this.searchResource.executeQuery(search.id(), null, searchUser));
-    }
-
-    @Test
-    public void failureToAddDefaultStreamsInSynchronousSearchLeadsTo403() {
-        final Search search = mockExistingSearch();
-
-        doThrow(new ForbiddenException()).when(search).addStreamsToQueriesWithoutStreams(any());
-
-        assertThatExceptionOfType(ForbiddenException.class)
-                .isThrownBy(() -> this.searchResource.executeSyncJob(search, 0, searchUser));
-    }
-
-    @Test
     public void guardExceptionOnPostLeadsTo403() {
         final Search search = mockNewSearch();
 
@@ -317,11 +135,6 @@ public class SearchResourceTest {
         final Search search = Search.builder().id(null).build();
 
         this.searchResource.createSearch(search, searchUser);
-    }
-
-    private void mockCurrentUserName(String name) {
-        when(currentUser.getName()).thenReturn(name);
-        when(searchUser.username()).thenReturn(name);
     }
 
     private void throwGuardExceptionFor(Search search) {
@@ -354,9 +167,6 @@ public class SearchResourceTest {
 
         final SearchJob searchJob = mock(SearchJob.class);
         when(searchJob.getResultFuture()).thenReturn(CompletableFuture.completedFuture(null));
-        when(searchJobService.create(any(), any())).thenReturn(searchJob);
-
-        when(queryEngine.execute(any())).thenAnswer(invocation -> invocation.getArgument(0));
 
         return search;
     }
